✅ EvalGen Beta Public Release (#335)
This is a WIP release, warts and all, so that people who want to try out EvalGen can finally try it out.

There is still polishing to do. However, this version of EvalGen works and is baseline functional.

This release also finally makes the VisNode able to plot on different x-axis if the evaluator outputs a dictionary (which the MultiEval does). 

It also bumps the genAI feature models for OpenAI to the gpt-4o line.